move aimmo worker simulation folder inside avatar_api folder
