+++ .travis.yml
@@ -90,10 +90,14 @@ jobs:
       before_install:
         - cd aimmo-game-worker
         - docker build --target tester -t ocadotechnology/aimmo-game-worker:test .
+        - pip install git+https://github.com/ocadotechnology/python-semantic-release.git@add_beta_releases
       install:
         - pip install codecov
       script:
         - docker run -it -p 5000:5000 ocadotechnology/aimmo-game-worker:test python setup.py test
+        - git config --global user.name "semantic-release (via TravisCI)"
+        - git config --global user.email "semantic-release@travis"
+        - semantic-release publish --branch=${TRAVIS_BRANCH} --build=${TRAVIS_BUILD_NUMBER} --dev=${DEPLOY_TO_DEV}
       after_success:
         # Get the coverage data from the container
         - read CONTAINERID <<< $(docker ps --all -q | head -1)
