+++ aimmo-game-worker/avatar_api/simulation/__init__.py
@@ -1,0 +1,1 @@
+__version__ = "0.43.0"
