+++ aimmo-game-worker/avatar_api/simulation/world_map.py
