+++ aimmo-game-worker/setup.cfg
@@ -1,0 +1,2 @@
+[semantic_release]
+version_variable = simulation/__init__.py:__version__
